Switched from KLModelObject to TTCoding.
Before this commit, both TTStyle and TTShape inherited from KLModelObject,
but now I have changed it so that they just inherit from NSObject
and are responsible for declaring that they conform to both the
NSCoding and NSCopying protocols.

I also added a TTCoding.h file that includes some convenient macros
for subclasses of TTStyle and TTShape to use when implementing the
NSCoding methods.